Test filtering aggregation modules without post-processing
Easy to reproduce with clinc_dummy by selecting the outcome=IncorrectAndRejected filter or the prediction=NO_INTENT filter. In those two examples, there should not be any correct predictions (green) in the distribution bars, nor in the confusion matrix. Filtering of all aggregation modules ignore `without_postprocessing`. It doesn't show in the confidence histogram since the "filtering"/graying out happens in the front end, but if I use the back end the same thing happens. It doesn't affect the utterance table as its filtering is not done in an aggregation module.