Return plain text errors
We used to expect error responses to have a JSON body like `{"detail": string}`. This was dangerous as the front end would fail if it encountered invalid JSON or plain text. And, for some obscure reason, we were still getting some errors in plain text, including some 500 Internal Server Error even though we had a handler for `HTTP_500_INTERNAL_SERVER_ERROR: handle_internal_error` that was returning some JSON. Also, the front end could receive network errors that don't come from the back end, for example if it is unable to reach the back end.

So, the easiest solution is to expect plain text from all error responses. To our knowledge, by default, only `HTTPException`s and `ValidationError`s used to return JSON, so I propose changing those handlers to plain text.

Worst case, if an error turns out to be JSON, it will still be valid text. The other way around is not true.