Remove nonfunctional 500 handler
Because of what appears to be a bug in FastAPI, our `HTTP_500_INTERNAL_SERVER_ERROR: handle_internal_error` supposed to handle unexpected errors as `{"detail": "Internal server error"}` works for some 500 errors, but not for others. Let's remove it for now, and treat all 500 errors as unexpected, by not showing them to the user.